+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes in 11.1
 - Prevent using App name, username, or email in the password in PasswordReset.php, FirstLogin.fnc.php, Inputs.php & Preferences.php
 - JS Add userInputs param to prevent using App name, username, or email in the password in jquery-passwordstrength.js
 - Return EMAIL column for students too (empty if "Student email field" not set) in User.fnc.php
+- Fix SQL error if delete Student email field, reset in StudentFields.php
 
 Changes in 11.0.2
 -----------------